requirements.txt
torch
torchvision
gradio
Pillow
gunicorn